[AST] Replace all occurences of ArraySeq
Most changed to List; some cases uses Map, when there is a suitable
alternative key (e.g. Wind).

The difference in efficiency of List vs ArraySeq is not much different,
since we either have very small collections (e.g. collections of tiles)
that will not grow very large anyways. For those large collections (e.g.
matches in a file, or frames in a match) it is not going to differ,
since we will traverse the whole collection for the most frequent
operations.

In anycase no one is supposed to create large paifu files with multiple
matches. Such large files are unfriendly to both human and computer.